properly handle session termination
A case was missed when the session is stopped determinsticially, but was
relying on the next underlying event to be emitted to handle setting the
session to `null'`.  Another asynchronous event could also happen in
between where '.stop()' was called, but the Session wasn't marked as
terminated already.

This change uses a cached Promise to resolve whether underlying session
is proplery stopped and having the wrappers internal state set
appropriately.