restructure type definition files and directories
This change introduces a level of indirection for referencing type
definitions and a convention for extending/overriding definitions
provided by the [DefinitelyTyped](http://http://definitelytyped.org/)
project.

The primary motiviations for introducing another level of indirection
for referencing type definitions are:

1. Provide each `.ts` file stability for the `.d.ts` file it references.
2. The subdirectory `typings`, where the stable `.d.ts` reference is
   located, also serves as the directortory to place
   extensions/overrides to third-party type definitions.

The consequence is that each `.ts` files only needs to reference
`typings/definitions.d.ts`, which will remain stable in event of file
moves.

Additionally, overrides may be placed in the `typings` subdirectory and
need only to be referenced by `typings/definitions.d.ts`.