Python 2.7.6 HTTP API Historical Data Example
This example is compatible with Python 2.7.6, which requires use of the
low-level httplib for SSL connections, since the SSLContext class is not
available in the ssl module. The requests module would be more ideal here,
but is a third-party module, so was avoided in effort to stick to core python.