Add subscription over long-polling
Server usage:
- `subscribe`: **POST** only, with JSON subscription data as request
body
- `poll`: **GET** only, with **pollid** specified as url arguments.
	- **pollid** is an integer that serve the function of data recovery
and data acknowledgement. The client start the first poll request with
any arbitrary integer. The client then increment the pollid by **1**
in their next poll request in order to acknowledge receiving the
previous data and is ready to receive new data. If for any reason the
client doesn't receive the previous data, it will make another poll
request with the previous pollid, and the server will resend the same
data as it did in the previous response. **Note**: Once the pollid get
incremented, there's no way going back to the previous value, since
the incrementation is a signal to the server to drop old data.
- `unsubscribe`: **POST** only, with JSON correlation ids as request
body, will only unsubscribe the specified subscriptions. The
correlationIds validation logic is the same as implemented in
`websocket`
- **NOTE**: The `subscription` feature only works in **HTTPS** mode,
since we are using `client certificate` to map user with their
session.

Some implementation notes:
- Add `map.ts` as generic map implementation that conforms to ES6 map
interface
- Add `session.ts` to create stateful **session** for our http server,
which holds user specific data(subscriptions and related).
- Add `session-store.ts` to hold `apiSession` and to handle session
expiration
- Add `api-session.ts` middleware to handle session creation and
retrieval. These middleware will only be executed for subscription
related route, not request/response route.
- Add `historical-buffer-manager.ts` to handle multi-buffer functionality
- Add two test examples in `test` directory.