Change URL scheme to be more consistent & RESTful.
The paths now reflect the operation type (i.e.
request/response or subscription), and any additional info
is in the query params or body.

Details of how the URLs changed:

- Subscribe, unsubscribe, and poll are now one path, and
  we distinguish between them based on POST vs. GET and
  query params.
- Requests use query params rather than having everything
  in the path.
- /request query params now match the BLPAPI values exactly
  rather than omitting the "Request" at the end of the type.

This change also adds a request validator that is used for
the new query params, but which also can be used for any
future additions to what's in the request.

Closes #59.