handling of 'error' event for subscriptions
NodeJS's `EventEmitter` treats emitting 'error' as a special case - when
there are no listeners, an `Error` object is thrown on behalf of the
`EventEmitter`.  Due to this behavior, an unhandled exception was being
thrown from `blpapi-wrapper.ts` that wasn't allowing the
`SessionTerminated` event being propagated properly to the users of the
API.

This change now has subscriptions listen for the 'error' event to
properly update the internal state of outstanding subscriptions.

Closes #39