Simplify onResponse to remove unneeded promise layers.
The wrapper's request function is already asynchronous, and
now that session handling is done earlier in middleware, we
can make requests a bit simpler. I also took the opportunity
to simplify how the response text is formed.

It should be possible to do the same sort of thing for
subscriptions but that will be a separate change.