extend ISocket abstraction to hide message structure details
This change extends `interface ISocket` so that the message structure is
handled by a base class implementation.  `SocketBaseImpl` was introduced
to be a partial abstract-base-class that handles most of the `ISocket`
interface, while stubbing out with exceptions the methods that must be
implemented by the parent class.  The `Socket.IO` and `ws`
implementations are modified to subclass `SocketBaseImpl` and the APIs
called in `websocket-handler.ts` are updated appropriately.