simplify error handling for mappings to underlying event messages
This change takes advantage of the
[blpapi-node](http://github.com/bloomberg/blpapi-node) throwing an
exception or emitting an error that will get handled naturally by the
top-level `catch` handler.

For request/response, an `assert` has been added to ensure that there is
no programmer error in keeping the request/response listener map
up-to-date.  Subscriptions does not need the check since the opening of
the service would fail beforehand; the dead code for subscriptions is
removed.