Add support for AuthenticationMode=APPLICATION_ONLY
This change adds a config option for api.authenticationAppName.
When this is set, sessions are created with authentication options
set to APPLICATION_ONLY mode with the specified app name, and
the default Identity for the session is authorized appropriately.
Other authentication modes are not yet supported.

This change also has a small refactor of doRequest to support
the third kind of request -- rather than passing isAuthRequest,
we now pass the actual function to perform the request.