Constrain TypeScript version to allow only for patch-level changes
Due to issue #152, his constrains TypeScript to >= 1.4.1, < 1.5.0.  It
turns out that from npm v1 => v2, there was a change in `semver` in how
caret (i.e., ^) handles preleases that changed.  Since it appears that
TypeScript has breaking changes between minor versions, we need to be
more restrictive on version constraints for `typescript`.

Closes #152